requirements.txt
neo4j
langchain
faiss-cpu
groq
langchain_groq
langchain_huggingface
langchain_community
fastapi
uvicorn
requests
spacy
pandas
transformers
python-dotenv